+++ spec/chef/recipes/gitlab-rails_spec.rb
@@ -480,6 +480,23 @@ describe 'gitlab::gitlab-rails' do
           }
         end
       end
+
+      context 'when migrated local files cleanup worker is configured' do
+        it 'sets the cron value' do
+          stub_gitlab_rb(gitlab_rails: { geo_migrated_local_files_clean_up_worker_cron: '1 2 3 4 5' })
+
+          expect(chef_run).to render_file(gitlab_yml_path)
+            .with_content(/geo_migrated_local_files_clean_up_worker:\s+cron:\s+"1 2 3 4 5"/)
+        end
+      end
+
+      context 'when migrated local files cleanup worker is not configured' do
+        it 'does not set the cron value' do
+          expect(chef_run).to render_file(gitlab_yml_path).with_content { |content|
+            expect(content).not_to include('geo_migrated_local_files_clean_up_worker')
+          }
+        end
+      end
     end
 
     context 'Scheduled Pipeline settings' do